Initial Commit
This code will read the CIE Spark PDF Pitch Decks placed in a Folder and will read each PDF using PDFPlumber, extract the Text and create a CSV with various columns required for processing the Pitch Deck.  It will also perform some data cleanup and correction activity.  Will finally call an LLM (as of now Gemini - Add your API Key) to summarize the Pitch Deck, Get an evaluation on a scale of 10 for the overall cohesion and narrative of the pitch deck.